+++ src/operations/inflation.ts
@@ -1,9 +1,9 @@
+import { setSourceAccount } from "../util/operations.js";
 import xdr from "../xdr.js";
 import {
   InflationOpts,
   InflationResult,
   OperationAttributes,
-  OperationClass,
 } from "./types.js";
 
 /**
@@ -12,13 +12,12 @@ import {
  * @param opts.source - The optional source account.
  */
 export function inflation(
-  this: OperationClass,
   opts: InflationOpts = {},
 ): xdr.Operation<InflationResult> {
   const opAttributes: OperationAttributes = {
     sourceAccount: null,
     body: xdr.OperationBody.inflation(),
   };
-  this.setSourceAccount(opAttributes, opts);
+  setSourceAccount(opAttributes, opts);
   return new xdr.Operation(opAttributes);
 }